【M】将错误处理抽成单独模块
