+++ app/__init__.py
@@ -39,4 +39,4 @@ migrate = Migrate(app,db)
 login = LoginManager(app)
 login.login_view = 'login'
 
-from app import routes
+from app import routes, errors
